requirements.txt
flask==2.0.1
werkzeug==2.0.1
scikit-learn==0.24.2
pandas==1.3.0
python-dotenv==0.19.0
requests==2.26.0
SQLAlchemy==1.4.23
python-telegram-bot==20.7
numpy==1.21.1
pytest==6.2.5
gunicorn==20.1.0
redis==3.5.3
selenium==4.1.0
beautifulsoup4==4.9.3
undetected-chromedriver==3.1.5
lxml==4.9.0
